Move useful skeleton and adapter tests to icu_provider_source
