+++ provider/source/src/datetime/symbols.rs
@@ -301,3 +301,166 @@ symbols_from!(
         midnight,
     },
 );
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use icu_locale_core::langid;
+
+    #[test]
+    fn test_adapter_months_numeric() {
+        let symbols: DataPayload<GregorianDateSymbolsV1Marker> = crate::provider::Baked
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_locale(&langid!("en").into()),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+        let neo_month_abbreviated: DataPayload<GregorianMonthNamesV1Marker> = symbols
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_marker_attributes_and_locale(
+                    DataMarkerAttributes::from_str_or_panic("3"),
+                    &"en".parse().unwrap(),
+                ),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+
+        assert_eq!(
+            format!("{neo_month_abbreviated:?}"),
+            "Linear([\"Jan\", \"Feb\", \"Mar\", \"Apr\", \"May\", \"Jun\", \"Jul\", \"Aug\", \"Sep\", \"Oct\", \"Nov\", \"Dec\"])"
+        );
+    }
+
+    #[test]
+    fn test_adapter_months_map() {
+        let symbols: DataPayload<HebrewDateSymbolsV1Marker> = crate::provider::Baked
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_locale(&langid!("en").into()),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+        let neo_month_abbreviated: DataPayload<HebrewMonthNamesV1Marker> = symbols
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_marker_attributes_and_locale(
+                    DataMarkerAttributes::from_str_or_panic("3"),
+                    &"en".parse().unwrap(),
+                ),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+
+        assert_eq!(
+            format!("{neo_month_abbreviated:?}"),
+            "LeapLinear([\"Tishri\", \"Heshvan\", \"Kislev\", \"Tevet\", \"Shevat\", \"Adar\", \"Nisan\", \"Iyar\", \"Sivan\", \"Tamuz\", \"Av\", \"Elul\", \"\", \"\", \"\", \"\", \"Adar I\", \"Adar II\", \"\", \"\", \"\", \"\", \"\", \"\"])"
+        );
+    }
+
+    #[test]
+    fn test_adapter_weekdays_abbreviated() {
+        let symbols: DataPayload<HebrewDateSymbolsV1Marker> = crate::provider::Baked
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_locale(&langid!("en").into()),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+        let neo_weekdays_abbreviated: DataPayload<WeekdayNamesV1Marker> = symbols
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_marker_attributes_and_locale(
+                    DataMarkerAttributes::from_str_or_panic("3"),
+                    &"en".parse().unwrap(),
+                ),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+
+        assert_eq!(
+            format!("{neo_weekdays_abbreviated:?}"),
+            "LinearNamesV1 { symbols: [\"Sun\", \"Mon\", \"Tue\", \"Wed\", \"Thu\", \"Fri\", \"Sat\"] }"
+        );
+    }
+
+    #[test]
+    fn test_adapter_weekdays_short() {
+        let symbols: DataPayload<HebrewDateSymbolsV1Marker> = crate::provider::Baked
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_locale(&langid!("en").into()),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+        let neo_weekdays_short: DataPayload<WeekdayNamesV1Marker> = symbols
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_marker_attributes_and_locale(
+                    DataMarkerAttributes::from_str_or_panic("6s"),
+                    &"en".parse().unwrap(),
+                ),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+
+        assert_eq!(
+            format!("{neo_weekdays_short:?}"),
+            "LinearNamesV1 { symbols: [\"Su\", \"Mo\", \"Tu\", \"We\", \"Th\", \"Fr\", \"Sa\"] }"
+        );
+    }
+
+    #[test]
+    fn test_adapter_eras() {
+        let symbols: DataPayload<GregorianDateSymbolsV1Marker> = crate::provider::Baked
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_locale(&langid!("en").into()),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+        let neo_eras_wide: DataPayload<GregorianYearNamesV1Marker> = symbols
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_marker_attributes_and_locale(
+                    DataMarkerAttributes::from_str_or_panic("4"),
+                    &"en".parse().unwrap(),
+                ),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+
+        assert_eq!(
+            format!("{neo_eras_wide:?}"),
+            "Eras(ZeroMap { keys: [\"bce\", \"ce\"], values: [\"Before Christ\", \"Anno Domini\"] })"
+        );
+    }
+
+    #[test]
+    fn test_adapter_dayperiods() {
+        let symbols: DataPayload<TimeSymbolsV1Marker> = crate::provider::Baked
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_locale(&langid!("en").into()),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+        let neo_dayperiods_abbreviated: DataPayload<DayPeriodNamesV1Marker> = symbols
+            .load(DataRequest {
+                id: DataIdentifierBorrowed::for_marker_attributes_and_locale(
+                    DataMarkerAttributes::from_str_or_panic("3s"),
+                    &"en".parse().unwrap(),
+                ),
+                ..Default::default()
+            })
+            .unwrap()
+            .payload;
+
+        assert_eq!(
+            format!("{neo_dayperiods_abbreviated:?}"),
+            "LinearNamesV1 { symbols: [\"AM\", \"PM\", \"noon\", \"midnight\"] }"
+        );
+    }
+}
+
